Querying pom.parent.version now, if SNAPSHOT is contained in version string, no deb package will be published to aptly repo, if not, deb package colreg-ui will be published in aptly snapshots repo.

+++ Jenkinsfile
@@ -10,9 +10,20 @@ node {
     sh "'${mvnHome}/bin/mvn' -U -Pdariah.deb -Dmaven.test.failure.ignore clean package"
   }
 
-  stage('Publish') {
-    sh "aptly repo add snapshots colreg-ui/target/*.deb"
-    sh "aptly publish update trusty"
-    sh "rm colreg-ui/target/*.deb"
+  stage('Publish') { 
+    def pom = readMavenPom file: 'colreg-ui/pom.xml'
+    def uiVersion = pom.parent.version
+    def snapshot = uiVersion.contains("SNAPSHOT")
+
+    if (!snapshot) {
+        echo "publishing deb package colreg-ui for non SNAPSHOT version ${uiVersion}"
+
+        sh "aptly repo add snapshots colreg-ui/target/*.deb"
+        sh "aptly publish update trusty"
+        sh "rm colreg-ui/target/*.deb"
+    }
+    else {
+        echo "deb package colreg-ui for SNAPSHOT version ${uiVersion} will not be published"
+    }
   }
 }
